frontend/requirements.txt
streamlit==1.22.0
requests==2.31.0
Pillow==9.5.0
protobuf==3.20.0